convert relatizeDate from jQuery to normal required library
